requirements.txt
black==25.1.0
click==8.1.8
Faker==36.1.1
findspark==2.0.1
iniconfig==2.0.0
joblib==1.4.2
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
py4j==0.10.9.7
pyarrow==19.0.1
pyspark==3.5.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
